+++ packages/monaco/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.6.7](http://github.com///compare/@codecademy/monaco@0.6.6...@codecademy/monaco@0.6.7) (2022-02-02)
+
+**Note:** Version bump only for package @codecademy/monaco
+
+
+
+
+
+### [0.6.6](http://github.com///compare/@codecademy/monaco@0.6.5...@codecademy/monaco@0.6.6) (2022-02-02)
+
+**Note:** Version bump only for package @codecademy/monaco
+
+
+
+
+
+### [0.6.5](http://github.com///compare/@codecademy/monaco@0.6.4...@codecademy/monaco@0.6.5) (2022-02-01)
+
+**Note:** Version bump only for package @codecademy/monaco
+
+
+
+
+
 ### [0.6.4](http://github.com///compare/@codecademy/monaco@0.6.3...@codecademy/monaco@0.6.4) (2022-01-31)
 
 **Note:** Version bump only for package @codecademy/monaco